refactor to a single scalar.h
