+++ tests/svs/quantization/scalar/scalar.cpp
@@ -18,7 +18,7 @@
 #include "svs/quantization/scalar/scalar.h"
 #include "svs/core/data/simple.h"
 #include "svs/lib/meta.h"
-#include "svs/quantization/scalar/impl/scalar_impl.h"
+#include "svs/quantization/scalar.h"
 
 #include "tests/svs/core/data/data.h"
 #include "tests/utils/test_dataset.h"
@@ -103,12 +103,10 @@ CATCH_TEST_CASE("Testing SQDataset", "[quantization][scalar]") {
         // We already verified the scale is calculated correctly above.
         float max_error = 0.5 * sq_dataset.get_scale();
 
-        // Now assert that the compression isn't off by more than one.
         auto delta = std::vector<float>(original.dimensions());
         for (size_t i = 0; i < original.size(); ++i) {
             auto datum = original.get_datum(i);
             auto sq_datum = sq_dataset.get_datum(i);
-            auto reloaded_sq_datum = reloaded_sq_dataset.get_datum(i);
 
             for (size_t j = 0; j < dims; ++j) {
                 float r = datum[j] - sq_datum[j];
